+++ mangadex_library/lib/jsonSearchCommands.dart
@@ -1,3 +1,0 @@
-import 'mangadex_library.dart';
-
-abstract class JsonUtils {}