+++ mangadex_library/lib/models/common/manga_status.dart
@@ -1,0 +1,6 @@
+enum MangaStatus {
+  ongoing,
+  completed,
+  haitus,
+  cancelled,
+}