+++ mangadex_library/lib/models/common/tag_modes.dart
@@ -1,0 +1,4 @@
+enum TagsMode {
+  Or,
+  And,
+}