+++ mangadex_library/test/mangadex_library_example.dart
@@ -8,8 +8,8 @@ void main() {
 void printFilenames() async {
   // this function, needs a mangadex account username and password supplied
   // to retrive login token
-  var username = 'USERNAME'; // Put your username here
-  var password = 'PASSWORD'; // Put your password here
+  var username = 'riksantra'; // Put your username here
+  var password = 'Sikkim123.'; // Put your password here
 
   //The line below uses the login function and takes in
   //two String parameters, username and password and returns
@@ -22,7 +22,8 @@ void printFilenames() async {
     // The token is used to access various sections and therefore it is recommended to be made accessible at all times.
 
     var searchData = await lib.search(
-        'oregairu'); //This is a search function that queries mangadex for the name of a manga
+        query:
+            'oregairu'); //This is a search function that queries mangadex for the name of a manga
     // it returns a Search class instance
     // For now, it searches for the Oregairu manga. You may replace the String value with your desired query.